+++ org.eclipse.winery.cli/src/main/java/org/eclipse/winery/cli/WineryCli.java
@@ -24,6 +24,7 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.Properties;
 import java.util.SortedSet;
+import java.util.stream.Stream;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
 
@@ -84,10 +85,11 @@ public class WineryCli {
 	private static final Logger LOGGER = LoggerFactory.getLogger(WineryCli.class);
 	private static final String ARTEFACT_BE = "artefact";
 
-	private enum Verbosity {
+	enum Verbosity {
 		OUTPUT_NUMBER_OF_TOSCA_COMPONENTS,
 		OUTPUT_CURRENT_TOSCA_COMPONENT_ID,
-		OUTPUT_ERRORS
+		OUTPUT_ERRORS,
+		NONE
 	}
 
 	public static void main(String[] args) throws ParseException {
@@ -339,7 +341,7 @@ public class WineryCli {
 	}
 
 	private static void checkId(List<String> res, EnumSet<Verbosity> verbosity, DefinitionsChildId id) {
-		checkNamespaceUri(res, verbosity, id, id.getNamespace().getDecoded());
+		checkNamespaceUri(res, verbosity, id);
 		checkNcname(res, verbosity, id, id.getXmlId().getDecoded());
 	}
 
@@ -352,11 +354,11 @@ public class WineryCli {
 		}
 	}
 
-	private static void checkNamespaceUri(List<String> res, EnumSet<Verbosity> verbosity, DefinitionsChildId id, String uriStr) {
+	static void checkNamespaceUri(List<String> res, EnumSet<Verbosity> verbosity, DefinitionsChildId id) {
 		Objects.requireNonNull(res);
 		Objects.requireNonNull(verbosity);
 		Objects.requireNonNull(id);
-		Objects.requireNonNull(uriStr);
+		String uriStr = id.getNamespace().getDecoded();
 		if (!uriStr.trim().equals(uriStr)) {
 			printAndAddError(res, verbosity, id, "Namespace starts or ends with white spaces");
 		}
@@ -382,7 +384,12 @@ public class WineryCli {
 		}
 		boolean namespaceUriContainsDifferentType = DefinitionsChildId.ALL_TOSCA_COMPONENT_ID_CLASSES.stream()
 			.filter(definitionsChildIdClass -> !definitionsChildIdClass.isAssignableFrom(id.getClass()))
-			.map(definitionsChildIdClass -> Util.getTypeForComponentId(definitionsChildIdClass).toLowerCase())
+			// we have the issue that nodetypeimplementation also contains nodetype
+			// we do the quick hack and check for plural s and /
+			.flatMap(definitionsChildIdClass -> {
+				final String lowerCaseIdClass = Util.getTypeForComponentId(definitionsChildIdClass).toLowerCase();
+				return Stream.of(lowerCaseIdClass + "s", lowerCaseIdClass + "/");
+			})
 			.anyMatch(definitionsChildName -> uriStr.contains(definitionsChildName));
 		if (namespaceUriContainsDifferentType) {
 			printAndAddError(res, verbosity, id, "Namespace URI contains tosca definitions name from other type. E.g., Namespace is ...servicetemplates..., but the type is an artifact template");