+++ org.eclipse.winery.repository.rest/src/main/java/org/eclipse/winery/repository/rest/resources/entitytemplates/PropertiesResource.java
@@ -18,6 +18,7 @@ import javax.ws.rs.Consumes;
 import javax.ws.rs.GET;
 import javax.ws.rs.PUT;
 import javax.ws.rs.Produces;
+import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
@@ -32,11 +33,11 @@ import org.eclipse.winery.repository.backend.RepositoryFactory;
 import org.eclipse.winery.repository.rest.RestUtils;
 import org.eclipse.winery.repository.rest.resources.AbstractComponentInstanceResource;
 
+import io.github.adr.embedded.ADR;
 import org.eclipse.jdt.annotation.NonNull;
 import org.eclipse.jdt.annotation.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.w3c.dom.Element;
 
 public class PropertiesResource {
 
@@ -57,7 +58,7 @@ public class PropertiesResource {
 
 	@PUT
 	@Consumes({MediaType.APPLICATION_XML, MediaType.TEXT_XML})
-	public Response setProperties(TEntityTemplate.Properties properties) {
+	public Response setProperties(@ADR(6) TEntityTemplate.Properties properties) {
 		this.template.setProperties(properties);
 		return RestUtils.persist(this.res);
 	}
@@ -95,7 +96,13 @@ public class PropertiesResource {
 					LOGGER.debug("XML properties expected, but none found. Returning empty JSON.");
 					return Response.ok().entity("{}").type(MediaType.APPLICATION_JSON).build();
 				}
-				return Response.ok().entity(Util.getXMLAsString((Element) any)).type(MediaType.TEXT_XML).build();
+				try {
+					@ADR(6)
+					Response response = Response.ok().entity(Util.getXMLAsString(TEntityTemplate.Properties.class, props)).type(MediaType.TEXT_XML).build();
+					return response;
+				} catch (Exception e) {
+					throw new WebApplicationException(e);
+				}
 			}
 		} else {
 			Properties properties;