+++ org.eclipse.winery.repository.ui/src/app/instance/sharedComponents/implementations/implementations.component.ts
@@ -18,6 +18,7 @@ import { ImplementationWithTypeAPIData } from './implementationWithTypeAPIData';
 import { ModalDirective } from 'ngx-bootstrap';
 import { Utils } from '../../../wineryUtils/utils';
 import { WineryRowData, WineryTableColumn } from '../../../wineryTableModule/wineryTable.component';
+import { ToscaTypes } from '../../../wineryInterfaces/enums';
 
 @Component({
     selector: 'winery-instance-implementations',
@@ -27,17 +28,18 @@ import { WineryRowData, WineryTableColumn } from '../../../wineryTableModule/win
 })
 export class ImplementationsComponent implements OnInit {
 
-    implementationData: ImplementationAPIData[];
+    implementationData: ImplementationAPIData[] = [];
     loading = true;
     selectedCell: any;
     newImplementation: ImplementationAPIData = new ImplementationAPIData('', '');
     elementToRemove: ImplementationAPIData;
-    nameOfElementToRemove = '';
-    selectedNamespace = '';
+    implementationOrTemplate: ToscaTypes;
+    templateOrImplementationsLink: string;
+    selectedNamespace: string;
     validatorObject: WineryValidatorObject;
     columns: Array<WineryTableColumn> = [
         { title: 'Namespace', name: 'namespace', sort: true },
-        { title: 'Name', name: 'displayname', sort: true },
+        { title: 'Name', name: 'displayName', sort: true },
     ];
     @ViewChild('confirmDeleteModal') confirmDeleteModal: ModalDirective;
     @ViewChild('addModal') addModal: ModalDirective;
@@ -45,11 +47,11 @@ export class ImplementationsComponent implements OnInit {
     constructor(private sharedData: InstanceService,
                 private service: ImplementationService,
                 private notificationService: WineryNotificationService) {
-        this.implementationData = [];
     }
 
     ngOnInit() {
         this.getImplementationData();
+        this.implementationOrTemplate = Utils.getImplementationOrTemplateOfType(this.sharedData.toscaComponent.toscaType);
     }
 
     // region ######## table methods ########
@@ -68,25 +70,18 @@ export class ImplementationsComponent implements OnInit {
     addNewImplementation(localname: string) {
         this.loading = true;
         const type = '{' + this.sharedData.toscaComponent.namespace + '}' + this.sharedData.toscaComponent.localName;
-        const resource = new ImplementationWithTypeAPIData(this.selectedNamespace,
-            localname,
-            type);
+        const resource = new ImplementationWithTypeAPIData(this.selectedNamespace, localname, type);
         this.service.postImplementation(resource).subscribe(
             data => this.handlePostResponse(data),
             error => this.handleError(error)
         );
     }
 
-    onRemoveClick(data: any) {
+    onRemoveClick(data: ImplementationAPIData) {
         if (isNullOrUndefined(data)) {
             return;
         } else {
-            this.elementToRemove = new ImplementationAPIData(data.namespace, data.localname);
-
-            const regex = /.*>(.*)<+/g;
-            const match = regex.exec(data.localname);
-            this.nameOfElementToRemove = match[1];
-
+            this.elementToRemove = data;
             this.confirmDeleteModal.show();
         }
     }
@@ -116,10 +111,10 @@ export class ImplementationsComponent implements OnInit {
     private handleData(impl: ImplementationAPIData[]) {
         this.implementationData = impl;
         this.implementationData = this.implementationData.map(item => {
-            const url = '/#/' + Utils.getImplementationOrTemplateOfType(this.sharedData.toscaComponent.toscaType)
+            const url = '/#/' + this.implementationOrTemplate
                 + '/' + encodeURIComponent(encodeURIComponent(item.namespace))
                 + '/' + item.localname;
-            item.displayname = '<a href="' + url + '">' + item.localname + '</a>';
+            item.displayName = '<a href="' + url + '">' + item.localname + '</a>';
             return item;
         });
         this.loading = false;
@@ -134,9 +129,9 @@ export class ImplementationsComponent implements OnInit {
         this.loading = false;
         if (data.ok) {
             this.getImplementationData();
-            this.notificationService.success('Created new Implementation');
+            this.notificationService.success('Created new ' + Utils.getToscaTypeNameFromToscaType(this.implementationOrTemplate));
         } else {
-            this.notificationService.error('Failed to create Implementation');
+            this.notificationService.error('Failed to create ' + Utils.getToscaTypeNameFromToscaType(this.implementationOrTemplate));
         }
     }
 
@@ -144,9 +139,9 @@ export class ImplementationsComponent implements OnInit {
         this.loading = false;
         if (data.ok) {
             this.getImplementationData();
-            this.notificationService.success('Deletion of Implementation Successful');
+            this.notificationService.success('Deletion of ' + Utils.getToscaTypeNameFromToscaType(this.implementationOrTemplate) + ' Successful');
         } else {
-            this.notificationService.error('Failed to delete Implementation failed');
+            this.notificationService.error('Failed to delete ' + Utils.getToscaTypeNameFromToscaType(this.implementationOrTemplate) + ' failed');
         }
     }