+++ org.eclipse.winery.repository.ui/src/app/wineryRepository.module.ts
@@ -39,7 +39,6 @@ import { RelationshipTypeImplementationModule } from './wineryMainModules/relati
 import { PolicyTemplateModule } from './wineryMainModules/policyTemplates/policyTemplate.module';
 import { ImportModule } from './wineryMainModules/imports/imports.module';
 import { WineryGitLogComponent } from './wineryGitLog/wineryGitLog.component';
-import { TemplatesOfTypeModule } from './instance/sharedComponents/templatesOfTypes/templatesOfType.Module';
 
 @NgModule({
     imports: [
@@ -53,7 +52,6 @@ import { TemplatesOfTypeModule } from './instance/sharedComponents/templatesOfTy
         SectionModule,
         WineryModalModule,
         TooltipModule.forRoot(),
-        TemplatesOfTypeModule,
         ServiceTemplateModule,
         NodeTypeModule,
         RelationshipTypeModule,