+++ CHANGELOG.md
@@ -15,6 +15,8 @@ This is similar to [Angular's CHANGELOG.md](https://github.com/angular/angular/b
 ## [unreleased]
 
 ### Changed
+- Add edit properties as XML
+- Fix lifecycle interface is now selected after click on "generate lifecycle interface"
 - Add available/not available indicator for LFS in git log component
 - Add ZIP-button for artifactemplate-files
 - Add license and readme support for all components